Feature/back cocktail

+++ api/src/main/java/com/cocktail27/api/service/UserService.java
@@ -1,7 +1,9 @@
 package com.cocktail27.api.service;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import com.cocktail27.api.dto.UserDTO;
 import com.cocktail27.api.model.User;
 import com.cocktail27.api.repository.UserRepository;
 
@@ -10,6 +12,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
+@Transactional
 public class UserService {
 
     @Autowired
@@ -27,11 +30,26 @@ public class UserService {
         return userRepository.findByUsername(username);
     }
 
-    public User updateUser(User user) {
-        return userRepository.save(user);
+    public User updateUser(UserDTO userDTO) throws RuntimeException {
+        Optional<User> optionalUser = userRepository.findById(userDTO.getId());
+        if (optionalUser.isPresent()) {
+            User user = optionalUser.get();
+            user.setUsername(userDTO.getUsername());
+            user.setName(userDTO.getName());
+            user.setEmail(userDTO.getEmail());
+            user.setRole(userDTO.getRole());
+            user.setImageUrl(userDTO.getImageUrl());
+            return userRepository.save(user);
+        }
+        throw new RuntimeException("User does not exist");
     }
 
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+    public void deleteUser(Long id) throws RuntimeException {
+        Optional<User> user = userRepository.findById(id);
+        if (user.isPresent()) {
+            userRepository.deleteById(id);
+            return;
+        }
+        throw new RuntimeException("User does not exist");
     }
 }
